Undo problematic bugfix in modules/core
The supposed bugfix introduced in commit `c7e50b3` actually induced the bug it described.
Special tests were performed and confirmed that the lastBlock+1 detection was handled correctly before the said bugfix. However, the fix itself intefered with the correct behaviour.
While the potential issue was acknowledged, the proposed fix was not neccessary.