Made ASM68K listing parser in CovnSym resistant to trailing CR bytes.
The parser of ASM68K output wrapper in ConvSym now handles trailing CR bytes properly. The problem with CR bytes (carriage return, a part of DOS-style line endings, CRLF byte sequence) occured on parsing raw ASM68K-generated listings with ConvSym on Unix systems, since fgets() treats CR as the part of the string, while on Windows (in text mode) CR is excluded from the resulting string. This caused symbol generation for the ErrorHandler module to fail, hence bundles cannot be built properly on Unix.

This commit also adds missing description for the /processLocals input option for the parser (options described in comments in the source code).