[modules/ErrorHandler] Fix a buffer overflow bug in "Formatter - Sym" that might've happened once in a moontime.
The buffer termination termination wasn't checked in-between writing the first and the second words of the displacement, which, if buffer was terminated by the flush function, could've caused writing past end of the buffer.

Instead of coordinating of writing two parts of the longword manually in "Formatter - Sym", an extra label was added in "Formatter - Hex" to allow displaying of a preemptively-swapped longword.