Error Handler, MDShell: Fix "symbol undefined" error in AS bundles.
Apparently AS cannot resolve symbols if a symbol forward-references a symbol that forward-references another symbol. There's only one level of depth for forward-reference resolution between passes.

The following example was breaking AS: `SOME_CONFIG: equ SomeSymbol`, where `SomeSymbol: equ Blob+$XX`. AS failed if `Blob` is a forward-reference. The solution is to move `Blob` above `SomeSymbol`, so it's no longer a forward-reference. Bundles implemented this by moving symbol table below the Blob for AS.