Core: Properly skip flags with arguments in KDebug.Write/.WriteLine.
This commit improves handling of flags with arguments (e.g. `setw,X`, `setx,X`), which are not supported in KDebug output. Previously, only the flag itself was ignored, regardless of arguments. However, this meant that if next byte is an argument, it would be printed. Now, flags with arguments ($F0..$FF) are properly detected and argument itself is skipped.