Core: Safer Console RAM pointer checks; don't lock-up if `usp` is corrupted or points to IO ports.
Console functions expect Console RAM pointer to be stored in `usp`. However, as a failsafe (in case someone tries to use them outside of initialized Console) each function also checked if RAM it pointed to is a valid Console RAM struct. This was done be examining a special byte of the struct, `Console.Magic`. If this byte has a value of `_ConsoleMagic` (`$5D`), we assume RAM pointer is valid.

This check is perfectly functional and failsafe even if `usp` points to invalid or unaligned ROM/RAM areas, with one exception: If `usp` points to IO ports, e.g. VDP area, attempts to read `Console.Magic` byte may lock up the system. Unfortunately, in advanced hacks and homebrew this may be the case as `usp` sometimes used to store VDP Port pointers to optimize Horizontal interrupt handlers.

This commit moves the magic value we check from Console RAM to MSB of `usp` itself. We know check MSB is a magic number to verify if it's a valid Console RAM pointer. Since pointers are 24-bit, upper 8-bits (MSB) were unused.

The new check is slightly slower than the original and uses more registers, but it shouldn't matter in the long run. I also deliberately avoided a few optimizations that would've hardcoded the expected `usp` value to MD's RAM range ($FF0000-$FFFFFF) so the code is more portable.