MD Debugger, MD Shell: Support projects with "ae+" assembly option in asm68k/axm68k bundles (not recommended).
"ae+" (automatic even enabled) is one of the more harmful assembly options in ASM68K that is turned on by default. It auto-aligns every "DC.B", which is likely to break formatted string generation (`__FSTRING_GenerateArgumentsCode` macro). Even though most of the projects turn it off by default, it's still a good idea to properly handle projects that don't.

This commit adds a failsafe in `__FSTRING_GenerateArgumentsCode` which temporarily disables "automatic even" if it was somehow turned on by default.